refactor: change type of the GroupBy parameter from string to interface{}
This change was made because SQL allows you to group by passing the column number, example "GROUP BY 3".
If we leave the parameter fixed as a string, the user is forced to pass "3" per parameter, but this way, changing it to interface{}, the user can pass 3 without quotes per parameter, which is more intuitive.